예외처리 및 리뷰등록  마무리

+++ server/src/controllers/reviewController.js
@@ -1,9 +1,17 @@
+import Book from "../models/book.js";
 import Review from "../models/review.js";
 import User from "../models/user.js";
 
 export const CReview = async (req, res) => {
   try {
-    const { content, authorId, bookId, rating } = req.body;
+    const { content, authorId, bookId, rating, ...otherData } = req.body;
+
+    if (Object.keys(otherData).length !== 0) {
+      return res.status(400).json({
+        success: false,
+        error: "유효하지 않은 데이터가 포함되어 있습니다.",
+      });
+    }
     if (content.trim() == "") {
       return res
         .status(422)
@@ -15,17 +23,23 @@ export const CReview = async (req, res) => {
         success: false,
       });
     }
+
     const isAuthor = await User.findById(authorId);
-    if (!isAuthor) {
-      return res
-        .status(400)
-        .json({ error: "유저를 찾을 수 없습니다", success: false });
-    }
     const isBookExists = await Book.findById(bookId);
+
+    if (!isAuthor) {
+      return res.status(400).json({
+        error: "리뷰를 등록할 유저를 찾을 수 없습니다",
+        success: false,
+      });
+    }
     if (!isBookExists) {
       return res
         .status(404)
-        .json({ error: "책을 찾을 수 없습니다", success: false });
+        .json({ error: "리뷰를 등록할 책을 찾을 수 없습니다", success: false });
+    }
+    if (rating !== undefined) {
+      reviewData.rating = rating;
     }
 
     const reviewData = {
@@ -34,18 +48,21 @@ export const CReview = async (req, res) => {
       book: bookId,
     };
 
-    if (rating !== undefined) {
-      reviewData.rating = rating;
-    }
-
     const review = new Review(reviewData);
-
     await review.save();
+
+    const user = await User.findById(authorId);
+    user.review.push(review._id);
+    await user.save();
+
+    const book = await Book.findById(bookId);
+    book.review.push(review._id);
+    await book.save();
 
     res.status(201).json({ data: review, success: true });
   } catch (error) {
     console.error("Error creating review: ", error);
-    res.status(500).json({ error: "Failed to create review", success: false });
+    res.status(500).json({ error: "리뷰 등록에 실패했습니다", success: false });
   }
 };
 
@@ -59,14 +76,21 @@ export const RReview = async (req, res) => {
     });
   } catch (error) {
     console.log(error);
-    res
-      .status(500)
-      .json({ error: "리뷰를 가져오는 중 에러 발생", success: false });
+    res.status(500).json({
+      error: "리뷰를 가져오는 중 에러가 발생했습니다",
+      success: false,
+    });
   }
 };
 
 export const UReview = async (req, res) => {
-  const { content, reviewId } = req.body;
+  const { content, reviewId, ...otherData } = req.body;
+  if (Object.keys(otherData).length !== 0) {
+    return res.status(400).json({
+      success: false,
+      error: "유효하지 않은 데이터가 포함되어 있습니다.",
+    });
+  }
   if (!content.trim() || !reviewId) {
     return res.status(422).json({
       error: "요청에 필요한 content 혹은 reviewId가 없습니다",
@@ -88,11 +112,17 @@ export const UReview = async (req, res) => {
     console.log(error);
     res
       .status(500)
-      .json({ error: "리뷰 업데이트 중 에러 발생", success: false });
+      .json({ error: "리뷰 업데이트 중 에러가 발생했습니다", success: false });
   }
 };
 export const DReview = async (req, res) => {
-  const { reviewId } = req.body;
+  const { reviewId, ...otherData } = req.body;
+  if (Object.keys(otherData).length !== 0) {
+    return res.status(400).json({
+      success: false,
+      error: "유효하지 않은 데이터가 포함되어 있습니다.",
+    });
+  }
   if (!reviewId) {
     return res.status(422).json({
       error: "요청에 필요한 reviewId가 없습니다",
@@ -100,10 +130,30 @@ export const DReview = async (req, res) => {
     });
   }
   try {
+    const review = await Review.findById(reviewId);
+    if (!review) {
+      return res.status(404).json({
+        error: "해당 리뷰를 찾을 수 없습니다",
+        success: false,
+      });
+    }
+
+    await User.updateOne(
+      { _id: review.author },
+      { $pull: { review: reviewId } }
+    );
+
+    await Book.updateOne({ _id: review.book }, { $pull: { review: reviewId } });
+
     await Review.findByIdAndDelete(reviewId);
-    res.status(200).json({ success: true });
+
+    res
+      .status(200)
+      .json({ success: true, message: "해당 리뷰가 삭제되었습니다" });
   } catch (error) {
     console.log(error);
-    res.status(500).json({ error: "리뷰 삭제 중 에러 발생", success: false });
+    res
+      .status(500)
+      .json({ error: "리뷰 삭제 중 에러가 발생했습니다", success: false });
   }
 };
